refactor: migrate to the bootstrap module
