+++ .gitignore
@@ -37,4 +37,4 @@ out/
 .vscode/
 
 # Bandlog specific
-bandlog-api/src/main/resources/application-secret.yaml
+application-secret.yaml
