--- bandlog-api/src/main/java/net/effize/bandlog/api/config/resolver/AuthUserParam.java
@@ -1,11 +1,0 @@
-package net.effize.bandlog.api.config.resolver;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface AuthUserParam {
-}
