+++ bandlog-bootstrap/src/main/resources/application.yaml
