+++ settings.gradle.kts
@@ -4,3 +4,4 @@ include("bandlog-api")
 include("bandlog-user")
 include("bandlog-shared")
 include("bandlog-team")
+include("bandlog-bootstrap")
